admin panelde değişiklikler yapıldı

+++ admin/referans.php
@@ -1,5 +1,10 @@
 
-<?php include 'include/header.php'; ?>
+<?php include 'include/header.php'; 
+$userId = $_SESSION["userId"]; 
+
+
+?>
+
 <section class="content-header">
         <h1>
         Admin Paneli
@@ -19,10 +24,11 @@
 
    right: 0px;
 
-   
+
 
    
 }
+.RbtnMargin { margin-left: 5px; }
             </style>
         </body>
     </html>
@@ -33,8 +39,19 @@
             <div class="row">
              <div class="box">
 <div class="box-header">Referanslar
-
-<button type="submit" class="btn btn-primary center" onclick="window.location.href='referans-ekle.php';"> Yeni Kayıt Ekle</button>
+<?php 
+   
+$yetki = DB::get("SELECT * FROM user_permissions WHERE userId='$userId' and permissionId = 1 ");
+if(count ($yetki)>0){
+?>
+<button type="submit" class="btn btn-primary pull-right RbtnMargin" onclick="window.location.href='referans-ekle.php';"> Yeni Kayıt Ekle</button>
+<?php } ?>
+<?php
+$id=$_SESSION['id'];
+$yetki = DB::get("SELECT * FROM users WHERE id='$id' and is_admin='admin'");
+if(count($yetki)>0) { ?>
+<button type="submit" class="btn btn-warning pull-right" onclick="window.location.href='referansonay.php';"> Onay Bekleyenler</button> 
+<?php } ?>
 </div>
 <div class="box-body">
 <table class="table table-sprited">
@@ -42,12 +59,13 @@
         <th>Referans ID</th>
         <th>Kategori ID</th>
         <th>Referans Adı</th>
+        <th>Resim </th>
         <th>İşlem</th>
     </thead>
     <tbody>
        <?php 
        
-       $calismalarim = DB::get("select * from referans");
+       $calismalarim = DB::get("select * from referans where durum=1");
        foreach($calismalarim as $row)
        {
            ?>
@@ -55,10 +73,22 @@
                <td><?=$row->referans_id?></td>
                <td><?=$row->kategori_id?></td>
                <td><?=$row->referansAd?></td>
+               <td><img src="referanslar/<?= $row->resim ?>" width=100>></td>
                <td>
                   <a href="referans-guncelle.php?referans_id=<?=$row->referans_id?>"><i class="fa fa-edit text-primary"></i></a>
+                  <?php
+         
+         $yetki = DB::get("SELECT * FROM user_permissions WHERE userId='$userId' and permissionId = 2 ");
+        
+         if(count ($yetki)>0){
+
+              ?>
                    <a href="?sil=<?=$row->referans_id?>"onclick="return confirm('Silmek istediğinize emin misiniz?');"><i class="fa fa-trash text-danger"></i></a>
-                  
+                   <?php
+                        }  
+                   
+                        ?>
+                     
                </td>
            </tr>
            <?php
@@ -76,9 +106,13 @@
 
 
     <?php 
+  
+  if(@$_GET["sil"])
+  {
+   
+   $yetki = DB::get("SELECT * FROM user_permissions WHERE userId='$userId' and permissionId = 2 ");
+   if(count($yetki)!=0){ 
 
-if(@$_GET["sil"])
-{
     $sil=DB::prepare("DELETE  FROM referans WHERE referans_id=:silinecekid");
     $sil->execute(["silinecekid"=> $_GET["sil"]]);
     if($sil)
@@ -89,10 +123,20 @@ if(@$_GET["sil"])
         echo "</script>";
         
     }
-}
 
 
-
+   }else{
+    echo "<script>
+    Swal.fire({
+       
+        icon: 'error',
+        title: 'Yetkisiz işlem',
+        showConfirmButton: false
+      
+      })
+    </script>";  
+   }
+  }
